requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
scikit-surprise
nltk
requests
tqdm
ipykernel
